Added list command + Fixed pack command
When packing as data, the folder architecture was wrong. When converting
from a zip to a dat, that still happens. Maybe adding a warning for it
would be nice.
When unpacking from data to dir, it would generate data_character_...
instead of recreating the folder structure. Now is behind a flag.